+++ src/apis/userRelatedAPI.js
@@ -1,4 +1,4 @@
-import axiosInstance from './config';
+import { axiosInstance } from './config';
 
 const client = axiosInstance;
